elish|enuma_defconfig: Change UCLAMP_BUCKETS_COUNT to 20
CONFIG_UCLAMP_BUCKETS_COUNT=20

Using the default setting 5 means the first bucket contains the
uclamp values from 0 to 19 (in percentage), in this case, if the
uclamp.min setting of a group is under 20, it will fall into this
bucket with other groups that have uclamp.min set to 0, which
increase the possibility of over boost. By setting the bucket
count to 20 will ease this situation, while a uclamp.min greater
than 4 will fall into a different bucket.

Bug: 170487162
Change-Id: Ie74349c2434c7739ed1fff9f26df71bb25c8ede4
Signed-off-by: Rick Yiu <rickyiu@google.com>